fix(coding-agent): harden live attach startup
