+++ packages/coding-agent/src/session/live-attach.ts
@@ -1,6 +1,6 @@
 import * as path from "node:path";
 import * as vcs from "@oh-my-pi/pi-natives/vcs";
-import { logger } from "@oh-my-pi/pi-utils";
+import * as logger from "@oh-my-pi/pi-utils/logger";
 import {
 	createDaemonBrokerClient,
 	createLiveSessionHost,
