+++ packages/utils/src/env.ts
@@ -2,6 +2,7 @@ import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
 import { getAgentDir, getConfigRootDir, getProjectDir, refreshDirsFromEnv } from "./dirs";
+import { isBunTestRuntime } from "./worker-host";
 
 export * from "./worker-host";
 
@@ -331,15 +332,6 @@ export function $envpos(name: string, defaultValue: number): number {
 	return parsed;
 }
 
-const BUN_TEST_ENTRY_PATTERN = /[._](?:test|spec)\.[cm]?[jt]sx?$/;
-
-/** True when the process is an explicitly marked test child or Bun is running a test entrypoint. */
-export function isBunTestRuntime(): boolean {
-	if (Bun.env.PI_TEST_RUNTIME === "1") return true;
-	const hasTestEnvironment = Bun.env.BUN_ENV === "test" || Bun.env.NODE_ENV === "test";
-	return hasTestEnvironment && BUN_TEST_ENTRY_PATTERN.test(Bun.main);
-}
-
 let terminalHeadless = isBunTestRuntime();
 
 /**
@@ -406,20 +398,6 @@ export function setInteractiveHost(interactive: boolean): boolean {
  */
 export function getDbBusyTimeoutMs(): number {
 	return isInteractiveHost() ? 5000 : 1000;
-}
-
-/**
- * True when this code is running inside a `bun build --compile` standalone
- * binary. Detects via the embedded virtual-filesystem path markers
- * (`$bunfs`, `~BUN`, or its URL-encoded form `%7EBUN`) in `import.meta.url`,
- * which Bun rewrites for every module bundled into the executable. The
- * `PI_COMPILED` env var (set by the build script's `--define`) is checked
- * first for cheap fast-path detection.
- */
-export function isCompiledBinary(): boolean {
-	if (process.env.PI_COMPILED || Bun.env.PI_COMPILED) return true;
-	const url = import.meta.url;
-	return url.includes("$bunfs") || url.includes("~BUN") || url.includes("%7EBUN");
 }
 
 const TRUTHY: Dict<boolean> = {
